Update torch_misuse.py with dtype

+++ torch_misuse.py
@@ -1,1 +1,1 @@
-x = torch. tensor([y])
+x = torch. tensor([y], dtype=torch.long)
